// src/app/shared/enums/type-message.enum.ts
export enum EnumTypeMessage {
  LOGIN_ERROR = 'LOGIN_ERROR',
  LOGIN_SUCCESS = 'LOGIN_SUCCESS',
  EXPIRED_TOKEN = 'EXPIRED_TOKEN',
  USER_LOGGED = 'USER_LOGGED',
  USER_NOT_LOGGED = 'USER_NOT_LOGGED',
  LOGOUT_SUCCESS = 'LOGOUT_SUCCESS',
  CREATE_SUCCESS = 'CREATE_SUCCESS',
  EDIT_SUCCESS =  'EDIT_SUCCESS',
  DELETE_SUCCESS = 'DELETE_SUCCESS',
  OPERATION_ERROR = 'OPERATION_ERROR',
  UNSELECTED_RECORD_PAGE_SAVE = 'UNSELECTED_RECORD_PAGE_SAVE'
}